namespace and version api endpoints
