+++ config/routes.rb
@@ -4,6 +4,9 @@ Rails.application.routes.draw do
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
   # Can be used by load balancers and uptime monitors to verify that the app is live.
   get "up" => "rails/health#show", as: :rails_health_check
-
-  resources :users, only: [:index, :create, :destroy]
+  namespace :api do
+    namespace :v1 do
+      resources :users, only: [:index, :create, :destroy]
+    end
+  end
 end
